requirements.txt
python-telegram-bot[job-queue]==22.6
pytz==2026.1.post1
